python-project/requirements.txt
avro==1.10.1
hazelcast-python-client==4.0
